refactor: move api call to services folder
